chore: extract example to component
