+++ apps/cookbook/src/app/examples/chart-example/chart-example.component.html
@@ -1,1 +1,1 @@
-<kirby-chart [height]="150" [options]="monthlyOverviewOptions"> </kirby-chart>
+<cookbook-monthly-overview-chart-example></cookbook-monthly-overview-chart-example>
