+++ apps/cookbook/src/app/examples/chart-example/examples/monthly-overview-chart-example/monthly-overview-chart-example.component.html
@@ -1,0 +1,1 @@
+<kirby-chart [height]="150" [options]="monthlyOverviewOptions"> </kirby-chart>
