+++ apps/cookbook/src/app/examples/examples.common.ts
@@ -48,6 +48,7 @@ import { ItemExampleComponent } from './item-example/item-example.component';
 import { DividerExampleComponent } from './divider-example/divider-example.component';
 import { ReorderListExampleComponent } from './reorder-list/reorder-list-example.component';
 import { DropdownExampleComponent } from '~/app/examples/dropdown-example/dropdown-example.component';
+import { MonthlyOverviewChartExampleComponent } from './chart-example/examples/monthly-overview-chart-example/monthly-overview-chart-example.component';
 
 // Example of "custom" icons
 export const iconSettings: IconSettings = {
@@ -77,6 +78,7 @@ export const COMPONENT_DECLARATIONS: any[] = [
   GridExampleComponent,
   AvatarExampleComponent,
   ChartExampleComponent,
+  MonthlyOverviewChartExampleComponent,
   FontsExampleComponent,
   SpinnerExampleComponent,
   ModalExampleComponent,
